test: break out Documents API unit tests into separate files
Restructures the test suite to mirror the implementation file structure, improving discoverability and maintainability.
